pyproject.toml
[project]
name = "flask-starter"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
  "flask>=3.1,<4.0",
  "gunicorn~=23.0.0",
  "python-dotenv>=1.0.0",
  "flask-sqlalchemy>=3.0.0",
  "sqlalchemy>=2.0.0"
]
